Moved authentication and registration to AuthController
